+++ src/p_setup.cpp
@@ -3933,7 +3933,7 @@ void P_SetupLevel (const char *lumpname, int position)
 	}
 	else reloop = true;
 
-	unsigned int startTime=0, endTime=0;
+	uint64_t startTime=0, endTime=0;
 
 	bool BuildGLNodes;
 	if (ForceNodeBuild)
@@ -3990,7 +3990,7 @@ void P_SetupLevel (const char *lumpname, int position)
 		// If the original nodes being loaded are not GL nodes they will be kept around for
 		// use in P_PointInSubsector to avoid problems with maps that depend on the specific
 		// nodes they were built with (P:AR E1M3 is a good example for a map where this is the case.)
-		reloop |= P_CheckNodes(map, BuildGLNodes, endTime - startTime);
+		reloop |= P_CheckNodes(map, BuildGLNodes, (uint32_t)(endTime - startTime));
 		hasglnodes = true;
 	}
 	else
